backend/requirements.txt
gunicorn==21.2.0
uvicorn==0.27.1
fastapi==0.109.2
python-dotenv==1.0.1
yt-dlp==2024.12.13
aiohttp==3.9.3
pydantic==2.6.1
jinja2==3.1.3
python-multipart==0.0.9
